feat: initial form implementation (#15)
The goal of creating the concept of a form is when a user enters a
command but doesn't supply all the required fields in order to
successfully run the command.  Instead of returning an error to the
user, we are dropping them into an interactive session where they can
step through all the required fields.

This commit is the initial implementation that has a form for org and
env fields.

Virtually all the commands we have require org, env, or both.  This
initial implementation should be able to scale with many more required
fields in a composable way.