feat: asset create form [sc-4311]
This provides a more fluid experience for creating an
asset.  The user will be presented with a series of questions that they
will answer in order to create an asset.

All of the questions except for the asset name are select dropdowns
which will ensure that the user enters all the inputs correctly.