refactor: feature based folder organization (#19)
The goal with this refactor is to keep things related to an entity (e.g.
asset, organization, environment, etc) colocated with each other.  This
has a few benefits:

- when refactoring an entity, all the code is located next to each other
- there is a home for all things related to an entity

I've also tried to flatten the folder structure so there is less
nesting.

All folders and files are singular by design.

I also tried to remove any possible naming collisions.  It's
very common for us to have a variable in our functions `assets` and
further we iterate over them and have an `asset` variable name.  This
often collides with any packages that we've named `asset` or `assets`.
Now we have a `libasset` package that is easy to autocomplete + import.

I also removed the `common` package collision by only having
`ui/common`.

We will want to continue to experiment with a feature based folder
organization, but I think this helps lay the foundation for our code
orgnaization.

The only caveat is that I left `cmd` untouched.  I felt like hiding
commands under the lib folder might be confusing.